refactor: Ajuste na estrutura de pacotes da aplicacao
