+++ app/Models/DTOs/PerfilTarefaDTO.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace App\Models;
+namespace App\Models\DTOs;
 
 use \Mviniis\ConnectionDatabase\DTO\DTO;
